+++ Admin.UI/Service/PvOpenMarketsService.cs
@@ -2,24 +2,35 @@
 using Admin.UI.Service.IService;
 using Newtonsoft.Json;
 using System.Net;
+using System.Net.Http.Headers;
 
 namespace Admin.UI.Service
 {
     public class PvOpenMarketsService : IPvOpenMarketsService
     {
         private readonly HttpClient _httpClient;
-
-        public PvOpenMarketsService(HttpClient httpClient)
+        private readonly IHttpContextAccessor _httpContextAccessor;
+        public PvOpenMarketsService(HttpClient httpClient, IHttpContextAccessor httpContextAccessor)
         {
 
             _httpClient = httpClient;
-
+            _httpContextAccessor = httpContextAccessor;
         }
 
         public async Task<PvOpenMarketsDto> DeleteOpenMarketAsync(int id)
         {
             try
             {
+                string token = _httpContextAccessor.HttpContext.Request.Cookies["Token"];
+
+                if (string.IsNullOrEmpty(token))
+                {
+                    // Handle case where token is missing or not retrieved properly
+                    throw new Exception("Token not found or invalid.");
+                }
+
+                // Set up HttpClient with the token in the Authorization header
+                _httpClient.DefaultRequestHeaders.Authorization = new AuthenticationHeaderValue("Bearer", token);
                 HttpResponseMessage response = await _httpClient.DeleteAsync($"https://localhost:7024/api/PvOpenMarketsAPI/Delete/{id}");
 
                 // Check if the response indicates a failure (non-success status code)
@@ -53,7 +64,19 @@ namespace Admin.UI.Service
         public async Task<PvOpenMarketsDto> GetOpenMarketByIdAsync(int id)
         {
             try
+
             {
+
+                string token = _httpContextAccessor.HttpContext.Request.Cookies["Token"];
+
+                if (string.IsNullOrEmpty(token))
+                {
+                    // Handle case where token is missing or not retrieved properly
+                    throw new Exception("Token not found or invalid.");
+                }
+
+                // Set up HttpClient with the token in the Authorization header
+                _httpClient.DefaultRequestHeaders.Authorization = new AuthenticationHeaderValue("Bearer", token);
                 HttpResponseMessage response = await _httpClient.GetAsync($"https://localhost:7024/api/PvOpenMarketsAPI/GetStateById{id}");
 
                 response.EnsureSuccessStatusCode();
@@ -75,6 +98,16 @@ namespace Admin.UI.Service
         {
             try
             {
+                string token = _httpContextAccessor.HttpContext.Request.Cookies["Token"];
+
+                if (string.IsNullOrEmpty(token))
+                {
+                    // Handle case where token is missing or not retrieved properly
+                    throw new Exception("Token not found or invalid.");
+                }
+
+                // Set up HttpClient with the token in the Authorization header
+                _httpClient.DefaultRequestHeaders.Authorization = new AuthenticationHeaderValue("Bearer", token);
                 HttpResponseMessage response = await _httpClient.GetAsync("https://localhost:7024/api/PvOpenMarketsAPI/GetAllAggregators");
 
                 response.EnsureSuccessStatusCode();
